+++ database/seeders/DatabaseSeeder.php
@@ -2,10 +2,7 @@
 
 namespace Database\Seeders;
 
-use Database\Seeders\TestOdkStuff\MediaTableSeeder;
-use Database\Seeders\TestOdkStuff\OdkProjectsTableSeeder;
-use Database\Seeders\TestTemplates\XlsformsTableSeeder;
-use Database\Seeders\TestTemplates\XlsformVersionsTableSeeder;
+use Database\Seeders\TestTemplates\DatasetsTableSeeder;
 use Illuminate\Database\Seeder;
 use Stats4sd\FilamentOdkLink\Database\Seeders\PlatformSeeder;
 
@@ -26,8 +23,7 @@ class DatabaseSeeder extends Seeder
         foreach (glob(database_path('seeders/Prep/*.php')) as $file) {
             $class = 'Database\\Seeders\\Prep\\' . pathinfo($file, PATHINFO_FILENAME);
             $this->call($class);
-            $this->call(DatasetsTableSeeder::class);
-    }
+        }
 
         // Call the test seeders locally
         if (app()->environment('local')) {
